[fix] fix the type mismatch

+++ src/main/java/com/yingying/distributed/hw2/distributedIO/cassandraIO/JSONAction.java
@@ -4,7 +4,7 @@ public class JSONAction extends CassandraAction {
 
     @Override
     public void close() {
-        for (byte item : buffer) {
+        for (int item : buffer) {
             session.execute("INSERT INTO number JSON '{\"pk\" : " + item + ", \"num\" : 1}';");
         }
     }
